@@ -1,17 +1,18 @@
 module github.com/kgantsov/dlock
 
-go 1.22.2
+go 1.24.1
 
 require (
 	github.com/ansrivas/fiberprometheus/v2 v2.6.1
 	github.com/danielgtaylor/huma/v2 v2.28.0
-	github.com/dgraph-io/badger/v4 v4.2.0
+	github.com/dgraph-io/badger/v4 v4.8.0
 	github.com/gofiber/fiber/v2 v2.52.5
-	github.com/hashicorp/go-msgpack/v2 v2.1.1
-	github.com/hashicorp/raft v1.6.1
-	github.com/rs/zerolog v1.33.0
-	github.com/sirupsen/logrus v1.4.2
-	github.com/stretchr/testify v1.9.0
+	github.com/hashicorp/go-msgpack/v2 v2.1.3
+	github.com/hashicorp/raft v1.7.3
+	github.com/kgantsov/raft-badgerstore v0.1.1
+	github.com/rs/zerolog v1.34.0
+	github.com/sirupsen/logrus v1.6.0
+	github.com/stretchr/testify v1.11.1
 	k8s.io/api v0.30.1
 	k8s.io/apimachinery v0.30.1
 	k8s.io/client-go v0.30.1
